requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-environ==0.9.0
django-filter==22.1
django-split-settings==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
importlib-metadata==4.12.0
Markdown==3.4.1
mysql-connector-python==8.0.30
numpy==1.23.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
zipp==3.8.1
